Renamed service folder to converter
